// components/Table/interfaces.tsx
export interface ColumnHeader {
    key: string;
    value: string; 
    type?: 'data' | 'date'; 
};